Move formatDuration into DatetimeUtil
Duration formatting is general time display, so it lives with the other
time formatters instead of in the grind display module.
